+++ src/plugin/filesystem.ts
